impl BranchPrecisionsHost
for the expected size of networks that this is built for, all precisions
together shouldn't take more than about 16 MB space, so they could
technically all be stored on device permanently.
Nevertheless, I want to store them on Host, because I believe that this
will make it easier ti implement multi-GPU versions of this software.