remove output of effect_sizes by default
these can be insanely huge, and probably slow down everything a lot as well, as there is a lot of writing going on.
There is no way to enable writing them in the cli at the moment.
Ideally, this should be an additional cli command that takes the
generated models, some data, and computes effect sizes on that, in a
format that does not take that much space (binary, and maybe not for
every single individuals, but some summary?)